reorganized imageseries unittests; added tests for compression/chunk; fixes to save.py
