+++ hexrd/imageseries/tests/test_process.py
@@ -1,0 +1,90 @@
+import numpy as np
+
+from .common import ImageSeriesTest, make_array, make_array_ims, compare
+
+from hexrd import imageseries
+from hexrd.imageseries import process, ImageSeries
+
+class TestImageSeriesProcess(ImageSeriesTest):
+
+    def _runfliptest(self, a, flip, aflip):
+        is_a = imageseries.open(None, 'array', data=a)
+        ops = [('flip', flip)]
+        is_p = process.ProcessedImageSeries(is_a, ops)
+        is_aflip = imageseries.open(None, 'array', data=aflip)
+        diff = compare(is_aflip, is_p)
+        msg = "flipped [%s] image series failed" % flip
+        self.assertAlmostEqual(diff, 0., msg=msg)
+
+    def test_process(self):
+        """Processed image series"""
+        is_a = make_array_ims()
+        is_p = process.ProcessedImageSeries(is_a, [])
+        diff = compare(is_a, is_p)
+        msg = "processed image series failed to reproduce original"
+        self.assertAlmostEqual(diff, 0., msg)
+
+    def test_process_flip_t(self):
+        """Processed image series: flip transpose"""
+        flip = 't'
+        a = make_array()
+        aflip = np.transpose(a, (0, 2, 1))
+        self._runfliptest(a, flip, aflip)
+
+    def test_process_flip_v(self):
+        """Processed image series: flip vertical"""
+        flip = 'v'
+        a = make_array()
+        aflip = a[:, :, ::-1]
+        self._runfliptest(a, flip, aflip)
+
+    def test_process_flip_h(self):
+        """Processed image series: flip horizontal"""
+        flip = 'h'
+        a = make_array()
+        aflip = a[:, ::-1, :]
+        self._runfliptest(a, flip, aflip)
+
+    def test_process_flip_vh(self):
+        """Processed image series: flip horizontal"""
+        flip = 'vh'
+        a = make_array()
+        aflip = a[:, ::-1, ::-1]
+        self._runfliptest(a, flip, aflip)
+
+    def test_process_flip_r90(self):
+        """Processed image series: flip horizontal"""
+        flip = 'ccw90'
+        a = make_array()
+        aflip = np.transpose(a, (0, 2, 1))[:, :, ::-1]
+        self._runfliptest(a, flip, aflip)
+
+    def test_process_flip_r270(self):
+        """Processed image series: flip horizontal"""
+        flip = 'cw90'
+        a = make_array()
+        aflip = np.transpose(a, (0, 2, 1))[:, ::-1, :]
+        self._runfliptest(a, flip, aflip)
+
+    def test_process_dark(self):
+        """Processed image series: dark image"""
+        a = make_array()
+        dark = np.ones_like(a[0])
+        is_a = imageseries.open(None, 'array', data=a)
+        apos = np.where(a >= 1, a-1, 0)
+        is_a1 = imageseries.open(None, 'array', data=apos)
+        ops = [('dark', dark)]
+        is_p = process.ProcessedImageSeries(is_a, ops)
+        diff = compare(is_a1, is_p)
+        self.assertAlmostEqual(diff, 0., msg="dark image failed")
+
+
+    def test_process_framelist(self):
+        a = make_array()
+        is_a = imageseries.open(None, 'array', data=a)
+        ops = []
+        frames = [0, 2]
+        is_p = process.ProcessedImageSeries(is_a, ops, frame_list=frames)
+        is_a2 = imageseries.open(None, 'array', data=a[tuple(frames), ...])
+        diff = compare(is_a2, is_p)
+        self.assertAlmostEqual(diff, 0., msg="frame list failed")
